requirements.txt
streamlit
tensorflow
opencv-python
mediapipe
numpy
pandas
matplotlib
scikit-learn
pillow
